Move asn logic to service
